refactor(daemon): split watcher convergence modules
